+++ front/pages/api/v1/w/[wId]/assistant/conversations/index.ts
@@ -16,8 +16,8 @@ import {
   createConversation,
   getConversation,
   normalizeContentFragmentType,
+  postNewContentFragment,
 } from "@app/lib/api/assistant/conversation";
-import { postNewContentFragment } from "@app/lib/api/assistant/conversation/content_fragment";
 import { postUserMessageWithPubSub } from "@app/lib/api/assistant/pubsub";
 import { Authenticator, getAPIKey } from "@app/lib/auth";
 import { apiError, withLogging } from "@app/logger/withlogging";
